+++ arkane/ess/molpro.py
@@ -201,7 +201,7 @@ class MolproLog(ESSAdapter):
             mass1, num1 = get_element_mass(atom1)
             mass.append(mass1)
             number.append(num1)
-        number = np.array(number, np.int)
+        number = np.array(number, int)
         mass = np.array(mass, float)
         coord = np.array(coord, float)
         if len(number) == 0 or len(coord) == 0 or len(mass) == 0:
