+++ arkane/ess/psi4.py
@@ -160,7 +160,7 @@ class Psi4Log(ESSAdapter):
             mass.append(mass_)
             number.append(num_)
         coord = np.array(coord, float)
-        number = np.array(number, np.int)
+        number = np.array(number, int)
         mass = np.array(mass, float)
         if any(len(param) == 0 for param in [number, coord, mass]):
             raise LogError(f"Unable to read atoms from Psi4 geometry output file {self.path}.")
